添加swagger-ui模块和单位测试

+++ src/main/java/com/mtx/sell/repository/ProductCategoryRepo.java
@@ -1,9 +1,10 @@
 package com.mtx.sell.repository;
 
 import com.mtx.sell.model.ProductCategory;
-        import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCategoryRepo  extends JpaRepository<ProductCategory, Long> {
+public interface  ProductCategoryRepo extends JpaRepository<ProductCategory, Integer>{
 
 }
